refactor: move startup to desktop
