+++ desktop/src/cn/harryh/arkpets/controllers/SettingsModule.java
@@ -9,7 +9,7 @@ import cn.harryh.arkpets.Const;
 import cn.harryh.arkpets.guitasks.CheckAppUpdateTask;
 import cn.harryh.arkpets.guitasks.CheckEnvironmentTask;
 import cn.harryh.arkpets.guitasks.GuiTask;
-import cn.harryh.arkpets.platform.StartupConfig;
+import cn.harryh.arkpets.startup.StartupConfig;
 import cn.harryh.arkpets.platform.WindowSystem;
 import cn.harryh.arkpets.envchecker.EnvCheckTask;
 import cn.harryh.arkpets.utils.*;
